Fix fow canvas for devicePixelRatio
